+++ anaconda-project.yml
@@ -26,16 +26,18 @@ description: Implementation of DDPMs in Pytorch to understand how they work
 #
 commands:
   train:
-    unix: python ddpm_pytorch/train.py
-    env_spec: default
-  train-gpu:
     unix: python ddpm_pytorch/train.py accelerator=gpu devices=1
     env_spec: default
-
-
+  train-cpu:
+    unix: python ddpm_pytorch/train.py accelerator=cpu devices=1
+    env_spec: default
   test:
     unix: pytest tests
     env_spec: default
+  generate:
+    unix: python ddpm_pytorch/generate.py
+    env_spec: default
+
 variables:
   PYTHONPATH: ./
   HYDRA_FULL_ERROR: 1
@@ -99,11 +101,11 @@ env_specs:
     - linux-64
     # - osx-64  # remove cudatoolkit to support mac
     - win-64
-  pytorch-cpu:
-    packages:
-    - cpuonly
-    channels:
-    - pytorch
-    platforms:
-    - linux-64
-    - win-64
+#  pytorch-cpu:
+#    packages:
+#    - cpuonly
+#    channels:
+#    - pytorch
+#    platforms:
+#    - linux-64
+#    - win-64
